requirements.txt
Django==1.8.16
django-bootstrap-form==3.2.1
django-filter==0.11.0
djangorestframework==3.3.1
pika
pip
PyMySQL
requests
simplejson
